[Chore] fix placeDropdown component name
